Add assertions rather than just throwing errors
This took a while, and I wrote an assert method before I realized
that there was one that was built into scheme. I used that instead
and it seems to be working pretty well now.

I left in all of the places where I set the root to the parent
directory, because this makes it way easier to load in things.